backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow
python-multipart